Updated authentication and added logout.
Updated authenticationService by adding a new function which gets the current userData if there is some.  Updated the authentication middleware which now attaches the user data to the req object if there is one and attaches isAuthenticated based on the userData we get from the authentication service. Also added user logout.